pyproject.toml
[tool.poetry]
name = "podcast-qa"
version = "0.1.0"
description = "Your personal Podcast assistant"
authors = ["Luke Skyward <luke.skyward.ai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
pandas = "^2.0.1"
python-youtube = "^0.9.0"
google-api-python-client = "^2.86.0"
pydantic = "^1.10.7"
youtube-transcript-api = "^0.6.0"
langchain = "^0.0.156"
openai = "^0.27.6"
tiktoken = "^0.3.3"
scikit-learn = "^1.2.2"
matplotlib = "^3.7.1"
nltk = "^3.8.1"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.18"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
pinecone-client = "^2.2.2"
sentence-transformers = "^2.2.2"
llvmlite = "0.40.1"
markdown = "^3.4.4"
streamlit = "^1.25.0"
torch = ">=2.0.0, !=2.0.1"
vecs = "^0.3.0"
mangum = "^0.17.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
